requirements.txt
google-generativeai
python-dotenv==1.0.0
langchain==0.0.350
pillow==10.1.0
duckduckgo-search==4.1.0
numexpr==2.8.8